+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: Replace 2 gwei priority fee minimum non-negative minimum
+
 ## 2.17.1 (2023-12-07)
 
 - added: Network adaptor to zkSync for transaction querying